+++ src/ThunderEgg/Schur/Interface.h
@@ -41,7 +41,7 @@ namespace Schur
  *
  * @tparam D the number of cartesian dimensions on a patch.
  */
-template <size_t D> class Interface : public Serializable
+template <int D> class Interface : public Serializable
 {
 	public:
 	/**
